the last admin cannot delete itself

+++ backend/api/users.py
@@ -127,7 +127,11 @@ def delete_user(mail):
     if not user:
         raise ApiException("user not found", 404)
     if is_permitted(current_user, user):
-        g.user_datastore.delete_user(user)
+        for usr in g.user_datestore.user_model.objects:
+            if usr.has_role('admin') and usr['email'] != user['email']:
+                g.user_datastore.delete_user(user)
+        raise ApiException("You are the last surviving admin, you cannot delete yourself", 9001)
+
     else:
         make_response("Permission denied!", 403)
